Move kmip plugin to parts
